requirements.txt
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
lightgbm>=4.0.0
sentence-transformers>=2.0.0
transformers>=4.20.0
torch>=1.12.0
timm>=0.6.0
opencv-python>=4.5.0
Pillow>=8.0.0
requests>=2.25.0
tqdm>=4.60.0
albumentations>=1.0.0
optuna>=3.0.0
joblib>=1.1.0
scipy>=1.7.0
pyarrow>=5.0.0
